requirements.txt
deemix
fairseq==0.12.2
faiss-cpu==1.7.3
ffmpeg-python>=0.2.0
gradio==3.39.0
lib==4.0.0
librosa==0.9.1
numpy==1.23.5
onnxruntime_gpu
praat-parselmouth>=0.4.2
pedalboard==0.7.7
pydub==0.25.1
pyworld==0.3.4
Requests==2.31.0
scipy==1.11.1
soundfile==0.12.1
torchcrepe==0.0.20
tqdm==4.65.0
yt_dlp==2023.7.6
sox==1.4.1

# my custom
fastapi
uvicorn
python_dotenv
